refactor: 게시물 페이지 css 및 컴포넌트 리팩토링
