+++ src/pages/PostListPage/PostListPage.style.ts
@@ -1,0 +1,17 @@
+import { css } from '@emotion/react';
+
+export const headingWrapper = css({
+  display: 'flex',
+  justifyContent: 'space-between',
+  width: '80%',
+  margin: '0 auto',
+  alignItems: 'center',
+});
+
+export const postListWrapper = css({
+  display: 'flex',
+  width: '90%',
+  justifyContent: 'space-between',
+  flexWrap: 'wrap',
+  margin: '0 auto',
+});
